+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project (attempts to) adhere to [Semantic Versioning](http://semver.org/).
 
+## [0.3.0] - 2018-07-03
+### Added
+- Sprofile now contains lots of stuff (including covT and stuff)
+- Added the 'lightRAM' option
+- Added the 'onlyPickle' option
+
 ## [0.2.3] - 2018-07-02
 ### Added
 - Added the production of the pickled class Sprofile